delete attachments on delete item

+++ Trainer/ItemsReview/Services/ItemsReviewsManager.cs
@@ -1,4 +1,6 @@
-﻿using FluentValidation;
+﻿using Attachments.Core.Interfaces;
+using Attachments.Core.Models;
+using FluentValidation;
 using ItemsReview.Extensions;
 using ItemsReview.Interfaces;
 using ItemsReview.Models;
@@ -10,8 +12,8 @@ using Shared.Core.Utilities.Enums;
 using Shared.Core.Utilities.Extensions;
 using Shared.Core.Utilities.Models;
 using System;
-using System.Net;
 using System.Linq;
+using System.Net;
 
 namespace ItemsReview.Services
 {
@@ -21,11 +23,14 @@ namespace ItemsReview.Services
         private readonly IValidator<ItemReviewDto> _validator;
         private readonly IRatingManager<ItemsForReview> _ratingManager;
         private readonly ILogger<ItemsReviewsManager> _logger;
-        public ItemsReviewsManager(IUnitOfWork unitOfWork, IValidator<ItemReviewDto> validator, IRatingManager<ItemsForReview> ratingManager, ILogger<ItemsReviewsManager> logger)
+        private readonly IAttachmentsManager _attachmentsManager;
+
+        public ItemsReviewsManager(IUnitOfWork unitOfWork, IValidator<ItemReviewDto> validator, IRatingManager<ItemsForReview> ratingManager, IAttachmentsManager attachmentsManager, ILogger<ItemsReviewsManager> logger)
         {
             _unitOfWork = unitOfWork;
             _validator = validator;
             _ratingManager = ratingManager;
+            _attachmentsManager = attachmentsManager;
             _logger = logger;
         }
         public ResultMessage GetAll(ItemsReviewFilter filter)
@@ -33,7 +38,7 @@ namespace ItemsReview.Services
             try
             {
                 PagedResult<ItemReviewDto> result = new PagedResult<ItemReviewDto>();
-                result = _unitOfWork.ItemsReviewsRepository.Get().ApplyFilter(filter).OrderBy(c=>c.Name).GetPaged(filter.PageNo, filter.PageSize).Adapt(result);
+                result = _unitOfWork.ItemsReviewsRepository.Get().ApplyFilter(filter).OrderBy(c => c.Name).GetPaged(filter.PageNo, filter.PageSize).Adapt(result);
 
                 return new ResultMessage()
                 {
@@ -63,8 +68,17 @@ namespace ItemsReview.Services
 
             try
             {
+                if (newItemDto.ProfilePictureFile != null)
+                    newItemDto.ProfilePicture = _attachmentsManager.Save(new SavedFileDto
+                    {
+                        attachmentType = AttachmentTypesEnum.Products,
+                        CanChangeName = true,
+                        File = newItemDto.ProfilePictureFile
+                    });
+
                 _unitOfWork.ItemsReviewsRepository.Insert(newItemDto.Adapt<ItemsForReview>());
                 _unitOfWork.Commit();
+
                 return new ResultMessage()
                 {
                     Status = HttpStatusCode.OK
@@ -136,6 +150,16 @@ namespace ItemsReview.Services
                 var oldItem = _unitOfWork.ItemsReviewsRepository.GetById(id);
                 if (oldItem != null)
                 {
+                    if (itemReviewDto.ProfilePictureFile != null)
+                    {
+                        itemReviewDto.ProfilePicture = _attachmentsManager.Save(new SavedFileDto
+                        {
+                            attachmentType = AttachmentTypesEnum.Products,
+                            CanChangeName = true,
+                            File = itemReviewDto.ProfilePictureFile
+                        });
+                    }
+
                     oldItem.Name = itemReviewDto.Name;
                     oldItem.ProfilePicture = itemReviewDto.ProfilePicture;
                     _unitOfWork.ItemsReviewsRepository.Update(oldItem);
@@ -168,7 +192,16 @@ namespace ItemsReview.Services
         {
             try
             {
+                var item = _unitOfWork.ItemsReviewsRepository.GetById(id);
+                if (item == null)
+                {
+                    return new ResultMessage()
+                    {
+                        Status = HttpStatusCode.NotFound
+                    };
+                }
                 _unitOfWork.ItemsReviewsRepository.Delete(id);
+                _attachmentsManager.Delete(item.ProfilePicture);
                 _unitOfWork.Commit();
                 return new ResultMessage()
                 {
